+++ xknx/telegram/address.py
@@ -15,21 +15,21 @@ The module supports all different writings of group addresses:
 """
 from __future__ import annotations
 
-from abc import ABC
+from abc import ABC, abstractmethod
 from enum import Enum
 from re import compile as re_compile
-from typing import ClassVar, Optional, Union
+from typing import ClassVar, Optional, TypeVar, Union
 
 from xknx.exceptions import CouldNotParseAddress
 
 # TODO: typing - remove need for Optional here
-GroupAddressableType = Optional[Union["GroupAddress", str, int, tuple[int, int]]]
-IndividualAddressableType = Optional[
-    Union["IndividualAddress", str, int, tuple[int, int]]
-]
+GroupAddressableType = Optional[Union["GroupAddress", str, int]]
+IndividualAddressableType = Optional[Union["IndividualAddress", str, int]]
 InternalGroupAddressableType = Union["InternalGroupAddress", str]
 DeviceAddressableType = Union[GroupAddressableType, InternalGroupAddressableType]
 DeviceGroupAddress = Union["GroupAddress", "InternalGroupAddress"]
+# py3.10 backwards compatibility - in py3.11 typing.Self is available
+Self = TypeVar("Self", bound="BaseAddress")
 
 
 def parse_device_group_address(
@@ -46,36 +46,29 @@ def parse_device_group_address(
         raise ex
 
 
-def address_tuple_to_int(address: tuple[int, int]) -> int:
-    """
-    Convert the tuple `address` to an integer.
-
-    Valid values inside the `address` tuple are:
-    * Positive Numbers between 0 and 255 (binary)
-    """
-    if (
-        any(not isinstance(byte, int) for byte in address)
-        or any(byte < 0 for byte in address)
-        or any(byte > 255 for byte in address)
-    ):
-        raise CouldNotParseAddress(address)
-    return int(address[0] * 256 + address[1])
-
-
 class BaseAddress(ABC):
     """Base class for all knx address types."""
 
-    def __init__(self) -> None:
-        """Initialize instance variables needed by all subclasses."""
-        self.raw: int = 0
+    raw: int
 
-    def to_knx(self) -> tuple[int, int]:
+    @abstractmethod
+    def __init__(
+        self, address: IndividualAddressableType | GroupAddressableType
+    ) -> None:
+        """Initialize Address instance. To be implemented in derived class."""
+
+    @classmethod
+    def from_knx(cls: type[Self], raw: bytes) -> Self:
+        """Parse/deserialize from KNX/IP raw data."""
+        return cls(int.from_bytes(raw, "big"))
+
+    def to_knx(self) -> bytes:
         """
         Serialize to KNX/IP raw data.
 
-        Returns a 2-Byte tuple generated from the raw Value.
+        Returns a bytes object with length of 2 from the raw value.
         """
-        return (self.raw >> 8) & 255, self.raw & 255
+        return int.to_bytes(self.raw, 2, "big")
 
     def __eq__(self, other: object | None) -> bool:
         """
@@ -103,24 +96,21 @@ class IndividualAddress(BaseAddress):
 
     def __init__(self, address: IndividualAddressableType) -> None:
         """Initialize IndividualAddress class."""
-        super().__init__()
-        if isinstance(address, IndividualAddress):
+        if isinstance(address, int):
+            self.raw = address
+        elif isinstance(address, IndividualAddress):
             self.raw = address.raw
         elif isinstance(address, str):
             if address.isdigit():
                 self.raw = int(address)
             else:
                 self.raw = self.__string_to_int(address)
-        elif isinstance(address, tuple) and len(address) == 2:
-            self.raw = address_tuple_to_int(address)
-        elif isinstance(address, int):
-            self.raw = address
         elif address is None:
             self.raw = 0
         else:
             raise CouldNotParseAddress(address)
 
-        if self.raw > 65535:
+        if not 0 <= self.raw <= 65535:
             raise CouldNotParseAddress(address)
 
     def __string_to_int(self, address: str) -> int:
@@ -210,25 +200,21 @@ class GroupAddress(BaseAddress):
 
     def __init__(self, address: GroupAddressableType) -> None:
         """Initialize GroupAddress class."""
-        super().__init__()
-
-        if isinstance(address, GroupAddress):
+        if isinstance(address, int):
+            self.raw = address
+        elif isinstance(address, GroupAddress):
             self.raw = address.raw
         elif isinstance(address, str):
             if address.isdigit():
                 self.raw = int(address)
             else:
                 self.raw = self.__string_to_int(address)
-        elif isinstance(address, tuple) and len(address) == 2:
-            self.raw = address_tuple_to_int(address)
-        elif isinstance(address, int):
-            self.raw = address
         elif address is None:
             self.raw = 0
         else:
             raise CouldNotParseAddress(address)
 
-        if self.raw > 65535:
+        if not 0 <= self.raw <= 65535:
             raise CouldNotParseAddress(address)
 
     def __string_to_int(self, address: str) -> int: